+++ doc/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+.. rubric:: Development version
+
+- Make :class:`~.SensorSet` more generic and move into :mod:`aiokatcp.sensor`
+  package. It no longer takes a list of connections; instead, one may register
+  callbacks to get notification of removals.
+
 .. rubric:: Version 0.4.4
 
 - Support Python 3.7